Add files via upload
Check out our Jupyter notebook to see how we built a cool tool that understands if movie reviews are thumbs-up or thumbs-down. It's like a mini-movie critic! Peek into the behind-the-scenes code, understand how we trained it, and see how well it predicts sentiments. The notebook is your friendly guide to the world of movie reviews and AI!